Support for sunxi hardware cursor
Hardware cursor is necessary because it is also visible on
top of sunxi disp layers, while software cursor is not.

FIXME: there is one minor problem with negative cursor
positions. The hardware does not support them, so such
positions are just set to 0 for now. In the future this
can be solved better by changing the cursor picture and
showing only the parts which are visible on screen.