DRI2: more informative messages for /var/log/Xorg.0.log
Explain that AIGLX is normally expected to fail and the users should
not really worry about it. Also provide a warning in the case if the
driver has been compiled without libUMP support (it could be that
the user actually wanted 3D acceleration, but just has not installed
all the needed dependencies).

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>