Enable G2D acceleration by default on sun4i hardware
With the fallback to CPU backend for unsupported blits and also
threshold for avoiding small blits, now G2D should always provide
best overall performance.

The users of recent versions of xf86-video-sunxifb are supposed
to also have a reasonably recent version of linux-sunxi kernel.
Which includes the following fix:
  https://github.com/linux-sunxi/linux-sunxi/commit/3d49345343a1535b

The users of old kernels are going to see screen corruption on
dragging windows and scrolling. They just should upgrade :)

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>