Added initial XV extension support for sunxi hardware
Proper layer sharing between XV and DRI2 still needs to be implemented.
Additionally we still need NEON and/or G2D "textured overlay" as a
fallback solution for the composited desktop (NEON optimized XV is going
to be useful for a wide range of ARM devices). A bit of performance
tuning is also necessary.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>