configure.ac: workaround libump/pthreads build issue
In some systems libump library is built without an explicit pthreads
dependency. As the issue has been already confirmed to affect both
sunxi and odroid users (and maybe the users of the other mali400
based hardware), it is easier to just workaround the problem locally.
Otherwise we would need to hunt down all the libump packagers and
beg for the fix.

More details are at https://github.com/ssvb/xf86-video-sunxifb/issues/11

Reported-by: Patrick Wood
Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>