test: new gles-rgb-cycle-demo for testing the correctness of DRI2
The test program cycles through 3 colors (red, green, blue), so
it is easier to see if we get the color change pattern wrong.
Also the X11 window title is updated to indicate the current
color information. If we have any problems with window
decorations handling, they are likely to be exposed.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>