DRI2: Added new "SwapbuffersWait" option for xorg.conf
When enabled, it tries to avoid tearing in OpenGL ES applications.
Works on sunxi hardware in the case if the hardware overlay (sunxi
disp layer) is used for a DRI2 window. The name of this option and
the description in the man page has been borrowed from intel and
radeon drivers.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>