DRI2: Ensure correct ordering of frames after window resize
In double buffer mode, explicitly mark the buffers as designated
for odd or even frame position when putting them into queue. And
when swapping the buffers, use these flags to re-synchronize if
it is necessary. This prevents problems after window resize (when
gles-rgb-cycle-demo could expose a mismatch between the color
name in the window title and the actual window color).

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>