Check if the kernel framebuffer driver returns errors on bad ioctls
When probing for the copyarea ioctl, we want to be sure that the
kernel just does not return 0 (success) for any unsupported ioctls.
The rockchip vendor kernels have been reported to have this issue.

In the case if the support for the Raspberry Pi specific copyarea
ioctl was detected by mistake, moving windows or scrolling was
broken.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>